Move the definition of XDG_CACHE_HOME into RepoInfo
We're not using it anywhere else so we move it as close as possible to
where we use it.
