+++ zypp-logic/zypp/repo/PackageProvider.h
@@ -26,9 +26,6 @@ namespace zypp
   ///////////////////////////////////////////////////////////////////
   namespace repo
   {
-    namespace env {
-    }
-
     ///////////////////////////////////////////////////////////////////
     /// \class PackageProviderPolicy
     /// \brief Policies and options for \ref PackageProvider
